Push out macro ports
